services/worker/requirements.txt
google-api-python-client==2.157.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
Pillow==10.4.0
ImageHash==4.3.1
opencv-python==4.10.0.84
insightface==0.7.3
sentence-transformers==3.0.1
faster-whisper==1.0.3
